Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -19,7 +19,7 @@ int main() {
             break;
         case '/':
             if (num2 != 0) {
-                printf("%.2f\n", (float)num1 / num2);
+                printf("%d\n", num1 / num2);
             } else {
                 printf("Error: Division by zero is not allowed.\n");
             }
